feat: refactor loader to generic function
