+++ React-Native/pages/pages/React/design-pattern/react-design/src/App.tsx
@@ -1,14 +1,24 @@
 import { UserInfo } from "./Components/Server/user-info";
-import { ResourceLoader } from "./Components/SourceLoader";
+import { DataSourceLoader } from "./Components/DataSourceLoader";
+import axios from "axios";
+
+const getDataFromServer = async (url: string) => {
+  const response = await axios.get(url);
+
+  return response.data;
+};
 
 function App() {
   return (
     <main style={{ width: "100%" }}>
-      <ResourceLoader resourcePath="dummyId" resourceName="user">
+      <DataSourceLoader
+        getData={() => getDataFromServer("someUrl")}
+        resourceName="user"
+      >
         <UserInfo
           user={{ name: "hey", age: 30, country: "ja", books: ["oh", "hey"] }}
         />
-      </ResourceLoader>
+      </DataSourceLoader>
     </main>
   );
 }
